Ensure DSA signatures are encoded according to spec (#115)
This change implements the raw encoding and decoding of DSA
signatures according to RFC4253 Section 6.6 instead of relying
on the internal representation in the dsa crate. See #114 for
additional detail.

This change also includes a test case, verifying that we
can verify the response from ssh-agent for a DSA signing request.