ssh-key: support additional SSH key algorithms (#136)
This introduces an additional "catch-all" key type to ssh-key to support additional SSH key algorithms, as described in #135.

Adds an `AlgorithmName` type for additional algorithm names. The syntax
for additional algorithm names is described in section 6 of RFC4251.

Adds a new `Algorithm::Other` variant for representing additional
algorithms.

Breaking changes: `Algorithm::as_str`, `Algorithm::as_certificate_str`
now return `&str` instead of `&static str`.

Adds the `Keypair::Other` and `KeyData::Other` variants for
storing the key material of keys that use a custom algorithm.

Adds the `OpaqueKeypair` and `OpaqueKeyData` types for representing keys
meant to be used with an algorithm unknown to this crate (e.g. custom
algorithms). They are said to be opaque, because the meaning of their
underlying byte representation is not specified.