ssh-encoding: add derive feature + extend derive functionallity (#348)
* ssh-derive: add support for all struct and enum types

Extend the derive macros to support all struct and enum types, not just
simple structs.

This also introduces some custom attributes to control the derived code.
Supported attributes:

  - `#[ssh(length_prefixed)]`: will use length-prefix encoding/decoding
    for the struct/enum as a whole, or per field, depending on where the
    attribute is placed.
  - `#[repr(u8)]`, `#[repr(u32)]` etc.: Enum discriminants will be
    encoded/decoded depending on the specified repr used.

* ssh-encoding: add error variant `InvalidDiscriminant`

This is used in derived `Decode` implementations for enums. It holds a
u128 because discriminants may be as large as u128/i128. Storing the
value isn't strictly necessary, but allows a better error message.

* ssh-encoding: add derive feature

Adds an optional feature that re-exports `ssh-derive` macros and exposes
a documentation module with examples. The examples also serve as tests.

Also adds tests for the derive feature in the form of declaring various
structs and enums and asserting that encoding/decoding works as expected.